requirements.txt
_libgcc_mutex==0.1
_openmp_mutex==4.5
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.2
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
audioread==3.0.1
binutils_impl_linux-64==2.40
blas==1.0
blinker==1.7.0
brotli-python==1.0.9
bzip2==1.0.8
c-ares==1.19.1
ca-certificates==2024.3.11
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
cmake==3.26.4
cn2an==0.5.22
contourpy==1.2.1
crcmod==1.7
cryptography==42.0.5
cuda-cudart==11.8.89
cuda-cupti==11.8.87
cuda-libraries==11.8.0
cuda-nvrtc==11.8.89
cuda-nvtx==11.8.86
cuda-runtime==11.8.0
cycler==0.12.1
datasets==2.18.0
decorator==5.1.1
dill==0.3.8
distance==0.1.3
einops==0.7.0
expat==2.6.2
fastapi==0.110.2
ffmpeg==4.2.2
ffmpeg-python==0.2.0
ffmpy==0.3.2
filelock==3.13.1
fonttools==4.51.0
freetype==2.12.1
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
g2p-en==2.1.0
gast==0.5.4
gcc==13.2.0
gcc_impl_linux-64==13.2.0
gitdb==4.0.11
gitpython==3.1.43
gmp==6.2.1
gmpy2==2.1.2
gnutls==3.6.15
gradio==4.27.0
gradio-client==0.15.1
gxx==13.2.0
gxx_impl_linux-64==13.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.4
importlib-metadata==7.1.0
importlib-resources==6.4.0
inflect==7.2.1
intel-openmp==2023.1.0
jieba-fast==0.53
jinja2==3.1.3
jmespath==0.10.0
joblib==1.4.0
jpeg==9e
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kernel-headers_linux-64==2.6.32
kiwisolver==1.4.5
krb5==1.20.1
lame==3.100
lazy-loader==0.4
lcms2==2.12
ld_impl_linux-64==2.40
lerc==3.0
libcublas==11.11.3.6
libcufft==10.9.0.58
libcufile==1.9.1.3
libcurand==10.3.5.147
libcurl==8.5.0
libcusolver==11.4.1.48
libcusparse==11.7.5.86
libdeflate==1.17
libedit==3.1.20230828
libev==4.33
libffi==3.4.4
libgcc-devel_linux-64==13.2.0
libgcc-ng==13.2.0
libgomp==13.2.0
libidn2==2.3.4
libjpeg-turbo==2.0.0
libnghttp2==1.57.0
libnpp==11.8.0.86
libnvjpeg==11.9.0.86
libopus==1.3.1
libpng==1.6.39
librosa==0.10.1
libsanitizer==13.2.0
libssh2==1.10.0
libstdcxx-devel_linux-64==13.2.0
libstdcxx-ng==13.2.0
libtasn1==4.19.0
libtiff==4.5.1
libunistring==0.9.10
libuuid==1.41.5
libuv==1.44.2
libvpx==1.7.0
libwebp-base==1.3.2
lightning-utilities==0.11.2
llvm-openmp==14.0.6
llvmlite==0.42.0
lz4-c==1.9.4
markdown-it-py==3.0.0
markupsafe==2.1.3
matplotlib==3.8.4
mdurl==0.1.2
mkl==2023.1.0
mkl-service==2.4.0
mkl_fft==1.3.8
mkl_random==1.2.4
modelscope==1.14.0
more-itertools==10.2.0
mpc==1.1.0
mpfr==4.0.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
ncurses==6.4
nettle==3.7.3
networkx==3.1
nltk==3.8.1
numba==0.59.1
numpy==1.26.4
numpy-base==1.26.4
openh264==2.1.1
openjpeg==2.4.0
openssl==3.0.13
orjson==3.10.1
oss2==2.18.4
packaging==24.0
pandas==2.2.2
pillow==10.2.0
pip==23.3.1
platformdirs==4.2.1
pooch==1.8.1
proces==0.1.7
protobuf==4.25.3
pyarrow==16.0.0
pyarrow-hotfix==0.6
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic-core==2.18.2
pydeck==0.8.1b0
pydub==0.25.1
pygments==2.17.2
pyopenjtalk==0.3.3
pyparsing==3.1.2
pypinyin==0.51.0
pysocks==1.7.1
python==3.11.9
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytorch==2.1.1
pytorch-cuda==11.8
pytorch-lightning==2.2.3
pytorch-mutex==1.0
pytz==2024.1
pyyaml==6.0.1
readline==8.2
referencing==0.34.0
regex==2024.4.16
requests==2.31.0
rhash==1.4.3
rich==13.7.1
rpds-py==0.18.0
ruff==0.4.1
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
semantic-version==2.10.0
setuptools==68.2.2
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
sqlite==3.41.2
starlette==0.37.2
streamlit==1.33.0
sympy==1.12
sysroot_linux-64==2.12
tbb==2021.8.0
tenacity==8.2.3
threadpoolctl==3.4.0
tk==8.6.12
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torchaudio==2.1.1
torchmetrics==1.3.2
torchtriton==2.1.0
torchvision==0.16.1
tornado==6.4
tqdm==4.66.2
transformers==4.40.1
typeguard==4.2.1
typer==0.12.3
typing-extensions==4.11.0
tzdata==2024.1
urllib3==2.1.0
uvicorn==0.29.0
watchdog==4.0.0
websockets==11.0.3
wheel==0.41.2
x264==1
xxhash==3.4.1
xz==5.4.6
yaml==0.2.5
yapf==0.40.2
yarl==1.9.4
zipp==3.18.1
zlib==1.2.13
zstd==1.5.5
sentencepiece==0.2.0